Componentize channel handshake components
