+++ crates/relayer-components/src/chain/traits/message_builders/mod.rs
@@ -1,4 +1,3 @@
-pub mod channel;
 pub mod connection;
 pub mod create_client;
 pub mod receive_packet;
