+++ crates/relayer-components/src/relay/impls/channel/open_confirm.rs
@@ -2,9 +2,8 @@ use cgp_core::async_trait;
 
 use crate::chain::traits::client::client_state::CanQueryClientState;
 use crate::chain::traits::components::chain_status_querier::CanQueryChainHeight;
-use crate::chain::traits::message_builders::channel::{
-    CanBuildChannelHandshakeMessages, CanBuildChannelHandshakePayloads,
-};
+use crate::chain::traits::components::channel_handshake_message_builder::CanBuildChannelHandshakeMessages;
+use crate::chain::traits::components::channel_handshake_payload_builder::CanBuildChannelHandshakePayloads;
 use crate::relay::traits::chains::HasRelayChains;
 use crate::relay::traits::channel::open_confirm::ChannelOpenConfirmRelayer;
 use crate::relay::traits::components::ibc_message_sender::{CanSendSingleIbcMessage, MainSink};
